# packages/api-client/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Dive Hub API",
    "version": "0.3.0",
    "description": "Sign in with POST /api/auth/sign-in/email (Better Auth); the session cookie authenticates every other call."
  },
...
  "paths": {
...
    "/api/imports": {
      "post": {
        "summary": "Upload a FIT file or a zip archive (e.g. Garmin \"Export Original\") for import",
        "responses": {
          "202": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "id",
                    "status",
                    "uploadName",
                    "createdAt",
                    "finishedAt",
                    "errorCode",
                    "error",
                    "outcome"
                  ],
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "status": {
                      "anyOf": [
                        {
                          "type": "string",
                          "enum": [
                            "pending"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "processing"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "done"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "failed"
                          ]
                        }
                      ]
                    },
                    "uploadName": {
                      "type": "string"
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "finishedAt": {
                      "anyOf": [
                        {
                          "type": "string",
                          "format": "date-time"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "errorCode": {
                      "anyOf": [
                        {
                          "enum": [
                            "unsupported_file",
                            "processing_failed"
                          ],
                          "description": "Why the Import failed; clients translate it"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "error": {
                      "anyOf": [
                        {
                          "type": "string",
                          "description": "English detail for processing_failed"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "outcome": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": [
                          "fileName",
                          "result"
                        ],
                        "properties": {
                          "fileName": {
                            "type": "string"
                          },
                          "result": {
                            "enum": [
                              "created",
                              "attached",
                              "updated",
                              "unchanged",
                              "duplicate-candidate",
                              "skipped",
                              "failed"
                            ]
                          },
                          "diveId": {
                            "type": "string"
                          },
                          "recordingId": {
                            "type": "string"
                          },
                          "reason": {
                            "enum": [
                              "no_fit_file",
                              "not_a_dive",
                              "not_your_diver",
                              "overlaps_several_dives",
                              "max_depth_differs",
                              "file_failed"
                            ]
                          },
                          "message": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "code",
                    "error"
                  ],
                  "properties": {
                    "code": {
                      "enum": [
                        "sign_in_required",
                        "admins_only",
                        "not_found",
                        "invalid_input",
                        "internal_error",
                        "upload_missing",
                        "upload_too_large",
                        "setup_done",
                        "setup_token_invalid",
                        "invitation_invalid",
                        "invitation_not_found",
                        "email_taken",
                        "reset_link_invalid",
                        "user_not_found",
                        "user_disabled",
                        "last_admin",
                        "not_yourself",
                        "confirmation_mismatch",
                        "session_not_found",
                        "import_not_found",
                        "dive_not_found",
                        "recording_not_found"
                      ],
                      "description": "Stable, machine-readable reason; clients translate it"
                    },
                    "error": {
                      "type": "string",
                      "description": "English description, for logs and scripts"
                    }
                  }
                }
              }
            }
          },
          "413": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "code",
                    "error"
                  ],
                  "properties": {
                    "code": {
                      "enum": [
                        "sign_in_required",
                        "admins_only",
                        "not_found",
                        "invalid_input",
                        "internal_error",
                        "upload_missing",
                        "upload_too_large",
                        "setup_done",
                        "setup_token_invalid",
                        "invitation_invalid",
                        "invitation_not_found",
                        "email_taken",
                        "reset_link_invalid",
                        "user_not_found",
                        "user_disabled",
                        "last_admin",
                        "not_yourself",
                        "confirmation_mismatch",
                        "session_not_found",
                        "import_not_found",
                        "dive_not_found",
                        "recording_not_found"
                      ],
                      "description": "Stable, machine-readable reason; clients translate it"
                    },
                    "error": {
                      "type": "string",
                      "description": "English description, for logs and scripts"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Recent Imports of the signed-in User",
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "required": [
                      "id",
                      "status",
                      "uploadName",
                      "createdAt",
                      "finishedAt",
                      "errorCode",
                      "error",
                      "outcome"
                    ],
                    "properties": {
                      "id": {
                        "type": "string"
                      },
                      "status": {
                        "anyOf": [
                          {
                            "type": "string",
                            "enum": [
                              "pending"
                            ]
                          },
                          {
                            "type": "string",
                            "enum": [
                              "processing"
                            ]
                          },
                          {
                            "type": "string",
                            "enum": [
                              "done"
                            ]
                          },
                          {
                            "type": "string",
                            "enum": [
                              "failed"
                            ]
                          }
                        ]
                      },
                      "uploadName": {
                        "type": "string"
                      },
                      "createdAt": {
                        "type": "string",
                        "format": "date-time"
                      },
                      "finishedAt": {
                        "anyOf": [
                          {
                            "type": "string",
                            "format": "date-time"
                          },
                          {
                            "type": "null"
                          }
                        ]
                      },
                      "errorCode": {
                        "anyOf": [
                          {
                            "enum": [
                              "unsupported_file",
                              "processing_failed"
                            ],
                            "description": "Why the Import failed; clients translate it"
                          },
                          {
                            "type": "null"
                          }
                        ]
                      },
                      "error": {
                        "anyOf": [
                          {
                            "type": "string",
                            "description": "English detail for processing_failed"
                          },
                          {
                            "type": "null"
                          }
                        ]
                      },
                      "outcome": {
                        "type": "array",
                        "items": {
                          "type": "object",
                          "required": [
                            "fileName",
                            "result"
                          ],
                          "properties": {
                            "fileName": {
                              "type": "string"
                            },
                            "result": {
                              "enum": [
                                "created",
                                "attached",
                                "updated",
                                "unchanged",
                                "duplicate-candidate",
                                "skipped",
                                "failed"
                              ]
                            },
                            "diveId": {
                              "type": "string"
                            },
                            "recordingId": {
                              "type": "string"
                            },
                            "reason": {
                              "enum": [
                                "no_fit_file",
                                "not_a_dive",
                                "not_your_diver",
                                "overlaps_several_dives",
                                "max_depth_differs",
                                "file_failed"
                              ]
                            },
                            "message": {
                              "type": "string"
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/imports/{id}": {
      "get": {
        "summary": "One Import with its outcome",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "id",
                    "status",
                    "uploadName",
                    "createdAt",
                    "finishedAt",
                    "errorCode",
                    "error",
                    "outcome"
                  ],
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "status": {
                      "anyOf": [
                        {
                          "type": "string",
                          "enum": [
                            "pending"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "processing"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "done"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "failed"
                          ]
                        }
                      ]
                    },
                    "uploadName": {
                      "type": "string"
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "finishedAt": {
                      "anyOf": [
                        {
                          "type": "string",
                          "format": "date-time"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "errorCode": {
                      "anyOf": [
                        {
                          "enum": [
                            "unsupported_file",
                            "processing_failed"
                          ],
                          "description": "Why the Import failed; clients translate it"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "error": {
                      "anyOf": [
                        {
                          "type": "string",
                          "description": "English detail for processing_failed"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "outcome": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": [
                          "fileName",
                          "result"
                        ],
                        "properties": {
                          "fileName": {
                            "type": "string"
                          },
                          "result": {
                            "enum": [
                              "created",
                              "attached",
                              "updated",
                              "unchanged",
                              "duplicate-candidate",
                              "skipped",
                              "failed"
                            ]
                          },
                          "diveId": {
                            "type": "string"
                          },
                          "recordingId": {
                            "type": "string"
                          },
                          "reason": {
                            "enum": [
                              "no_fit_file",
                              "not_a_dive",
                              "not_your_diver",
                              "overlaps_several_dives",
                              "max_depth_differs",
                              "file_failed"
                            ]
                          },
                          "message": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "code",
                    "error"
                  ],
                  "properties": {
                    "code": {
                      "enum": [
                        "sign_in_required",
                        "admins_only",
                        "not_found",
                        "invalid_input",
                        "internal_error",
                        "upload_missing",
                        "upload_too_large",
                        "setup_done",
                        "setup_token_invalid",
                        "invitation_invalid",
                        "invitation_not_found",
                        "email_taken",
                        "reset_link_invalid",
                        "user_not_found",
                        "user_disabled",
                        "last_admin",
                        "not_yourself",
                        "confirmation_mismatch",
                        "session_not_found",
                        "import_not_found",
                        "dive_not_found",
                        "recording_not_found"
                      ],
                      "description": "Stable, machine-readable reason; clients translate it"
                    },
                    "error": {
                      "type": "string",
                      "description": "English description, for logs and scripts"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}